token-cli: Move cmd test into a separate file
