+++ token/cli/src/main.rs
@@ -3170,14 +3170,13 @@ fn handle_tx(
 mod tests {
     use {
         super::*,
-        assert_cmd::prelude::*,
         solana_client::blockhash_query::Source,
         solana_sdk::{
             bpf_loader,
             signature::{Keypair, Signer},
         },
         solana_test_validator::{ProgramInfo, TestValidator, TestValidatorGenesis},
-        std::{path::PathBuf, process::Command},
+        std::path::PathBuf,
     };
 
     fn clone_keypair(keypair: &Keypair) -> Keypair {
@@ -3714,17 +3713,5 @@ mod tests {
         let ui_account = config.rpc_client.get_token_account(&aux).unwrap().unwrap();
         assert_eq!(ui_account.mint, token.to_string());
         assert_eq!(ui_account.owner, aux_string);
-    }
-
-    #[test]
-    fn invalid_config_will_cause_commands_to_fail() {
-        let mut cmd = Command::cargo_bin("spl-token").unwrap();
-        let args = &["address", "--config", "~/nonexistent/config.yml"];
-
-        cmd.args(args)
-            .assert()
-            .stderr("error: Could not find config file `~/nonexistent/config.yml`\n");
-
-        cmd.args(args).assert().code(1).failure();
     }
 }
